tranform historical inflows from weeks to days
