+++ pkg_copy_paste/__init__.py
@@ -7,3 +7,4 @@ import pandas as pd
 import numpy as np
 import time
 import pkg_copy_paste.ClassCopyPaste
+import pkg_copy_paste.plexos_crear_csv
